% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_spectral_data.R
\name{classify_spectral_data}
\alias{classify_spectral_data}
\title{Classify Spectral Data}
\usage{
classify_spectral_data(input, library, output, SizeClasses, sizelabels)
}
\arguments{
\item{input}{Path to the input file}

\item{SizeClasses}{An array with the defined size class bins}
}
\value{
A matrix of the infile
}
\description{
This function inspects a pre-loaded library, and then removes any particles
with a match lower than 0.8, and/or that have been rejected in the sorting
sheet. It relies on data being preprocesesd by the earlier steps.
}
