% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spec_lib.R
\name{import_spec_lib}
\alias{import_spec_lib}
\title{Import Spec Lib}
\usage{
import_spec_lib(lib1file, lib1sheet, lib2file, lib2sheet)
}
\arguments{
\item{lib1file}{Path to the first processed library as XLSX}

\item{lib1sheet}{The excel sheet name to pick the data from}

\item{lib2file}{Path to the second processed library as XLSX}

\item{lib2sheet}{The excel sheet name to pick the data from}
}
\value{
A data frame containing a single particle library
}
\description{
This function takes two excel spreadsheets with the spectral library output
that has already been processed and combines them into one library
It relies on data being preprocesesd by the earlier steps.
}
