% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spec_lib.R
\name{process_spec_lib}
\alias{process_spec_lib}
\title{Process Spec Lib}
\usage{
process_spec_lib(file)
}
\arguments{
\item{file}{Path to the raw spectral library export from clarity}
}
\value{
A data frame containing the required columns for a particle library
}
\description{
This function takes the raw CSV spectral library output from clarity
and transforms it into a data frame that can then be classified
It should then be exported to an XLSX if using import_spec_lib
}
