% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_spectral_data.R
\name{classify_spectral_data}
\alias{classify_spectral_data}
\title{Classify Spectral Data}
\usage{
classify_spectral_data(
  LDIRData,
  SpecLib,
  HitRate = 0.8,
  SizeClasses = c(0, 10, 50, 100, 200, 500, 900)
)
}
\arguments{
\item{LDIRData}{Data frame containing LDIR particle analysis data}

\item{SpecLib}{Data frame containing spectral family mappings}

\item{HitRate}{Quality rate lower bound as decimal}

\item{SizeClasses}{Array of integers for size binning}
}
\value{
Input dataset classified by polymer family and identified as plastic or not
}
\description{
This function inspects a pre-loaded library, and then removes any particles
with a match lower than 0.8, and/or that have been rejected in the sorting
sheet. It relies on data being preprocesesd by the earlier steps.
}
